bitbox02/keystore: remove the stub file
